(refactor): move index.js into src/
- standardize w/ the rest of my libraries and better parity between
  test/ and src/
