+++ webpack.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-  entry: './index.js',
+  entry: './src/index.js',
   output: {
     path: './build',
     filename: 'index.js',
